+++ sdk/nodejs/resource.ts
@@ -15,6 +15,7 @@
 import { util } from "protobufjs";
 import { ResourceError, RunError } from "./errors";
 import { all, Input, Inputs, interpolate, Output, output } from "./output";
+import { getStackResource } from "./runtime";
 import { readResource, registerResource, registerResourceOutputs } from "./runtime/resource";
 import { getProject, getStack } from "./runtime/settings";
 import * as utils from "./utils";
@@ -142,10 +143,21 @@ export abstract class Resource {
 
     /**
      * @internal
+     * A collection of transformations to apply as part of resource registration.
+     *
+     * Note: This is marked optional only because older versions of this library may not have had
+     * this property, and marking optional forces consumers of the property to defensively handle
+     * cases where they are passed "old" resources.
+     */
+    // tslint:disable-next-line:variable-name
+    __transformations?: ResourceTransformation[];
+
+    /**
+     * @internal
      * A list of aliases applied to this resource.
      *
      * Note: This is marked optional only because older versions of this library may not have had
-     * this property, and marking optional forces conumers of the property to defensively handle
+     * this property, and marking optional forces consumers of the property to defensively handle
      * cases where they are passed "old" resources.
      */
     // tslint:disable-next-line:variable-name
@@ -156,7 +168,7 @@ export abstract class Resource {
      * The name assigned to the resource at construction.
      *
      * Note: This is marked optional only because older versions of this library may not have had
-     * this property, and marking optional forces conumers of the property to defensively handle
+     * this property, and marking optional forces consumers of the property to defensively handle
      * cases where they are passed "old" resources.
      */
     // tslint:disable-next-line:variable-name
@@ -268,6 +280,10 @@ export abstract class Resource {
                 : convertToProvidersMap((<ComponentResourceOptions>opts).providers);
             this.__providers = { ...this.__providers, ...providers };
         }
+
+        // Combine transformations inherited from the parent with transformations provided in opts.
+        const parent = opts.parent || getStackResource() || { __transformations: undefined };
+        this.__transformations = [ ...(opts.transformations || []), ...(parent.__transformations || []) ];
 
         this.__protect = !!opts.protect;
 
@@ -466,6 +482,12 @@ export interface ResourceOptions {
      * An optional customTimeouts configuration block.
      */
     customTimeouts?: CustomTimeouts;
+    /**
+     * Optional list of transformations to apply to this resource during construction. The
+     * transformations are applied in order, and are applied prior to transformation applied to
+     * parents walking from the resource up to the stack.
+     */
+    transformations?: ResourceTransformation[];
 
     // !!! IMPORTANT !!! If you add a new field to this type, make sure to add test that verifies
     // that mergeOptions works properly for it.
@@ -484,6 +506,58 @@ export interface CustomTimeouts {
      * The optional delete timeout represented as a string e.g. 5m, 40s, 1d.
      */
     delete?: string;
+}
+
+/**
+ * ResourceTransformation is the callback signature for the `transformations` resource option.  A
+ * transformation is passed the same set of inputs provided to the `Resource` constructor, and can
+ * optionally return back alternate values for the `props` and/or `opts` prior to the resource
+ * actually being created.  The effect will be as though those props and opts were passed in place
+ * of the original call to the `Resource` constructor.  If the transformation returns undefined,
+ * this indicates that the resource will not be transformed.
+ */
+export type ResourceTransformation = (args: ResourceTransformationArgs) => ResourceTransformationResult | undefined;
+
+/**
+ * ResourceTransformationArgs is the argument bag passed to a resource transformation.
+ */
+export interface ResourceTransformationArgs {
+    /**
+     * The Resource instance that is being transformed.
+     */
+    resource: Resource;
+    /**
+     * The type of the Resource.
+     */
+    type: string;
+    /**
+     * The name of the Resource.
+     */
+    name: string;
+    /**
+     * The original properties passed to the Resource constructor.
+     */
+    props: Inputs;
+    /**
+     * The original resource options passed to the Resource constructor.
+     */
+    opts: ResourceOptions;
+}
+
+/**
+ * ResourceTransformationResult is the result that must be returned by a resource transformation
+ * callback.  It includes new values to use for the `props` and `opts` of the `Resource` in place of
+ * the originally provided values.
+ */
+export interface ResourceTransformationResult {
+    /**
+     * The new properties to use in place of the original `props`
+     */
+    props: Inputs;
+    /**
+     * The new resource options to use in place of the original `opts`
+     */
+    opts: ResourceOptions;
 }
 
 /**