+++ CHANGELOG.md
@@ -18,6 +18,9 @@ CHANGELOG
   [#3183](https://github.com/pulumi/pulumi/pull/3183)
 - Add Codefresh CI detection.
 - Add `-c` (config array) flag to the `preview` command.
+- Adds the ability to provide transformations to modify the properties and resource options that
+  will be used for any child resource of a component or stack.
+  [#3174](https://github.com/pulumi/pulumi/pull/3174)
 
 ## 1.1.0 (2019-09-11)
 
@@ -48,6 +51,7 @@ CHANGELOG
 
 - Filter the list of templates shown by default during `pulumi new`.
   [#3147](https://github.com/pulumi/pulumi/pull/3147)
+
 ## 1.0.0-beta.4 (2019-08-22)
 
 - Fix a crash when using StackReference from the `1.0.0-beta.3` version of